Initial setup. Setup tableView.
