+++ SettingsScreenApplication/Resources/Assets.xcassets/AccentColor.colorset/Contents.json
