+++ SettingsScreenApplication/Resources/Assets.xcassets/Contents.json
