+++ SettingsScreenApplication/Sources/ViewController.swift
@@ -1,0 +1,47 @@
+//
+//  ViewController.swift
+//  SettingsScreenApplication
+//
+//  Created by Nika Semenkova on 01.02.2023.
+//
+
+import UIKit
+
+class ViewController: UIViewController {
+    
+    // MARK: UI Elements
+    
+    private lazy var tableView: UITableView = {
+        let tableView = UITableView(frame: .zero, style: .grouped)
+        
+        tableView.translatesAutoresizingMaskIntoConstraints = false
+        return tableView
+    }()
+    
+    // MARK: Lifecycle
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+    }
+
+    // MARK: Setups
+    private func setupView() {
+        view.backgroundColor = .white
+        title = "Настройки"
+        navigationController?.navigationBar.prefersLargeTitles = true
+    }
+    
+    
+    private func setupHierarchy() {
+        view.addSubview(tableView)
+    }
+    
+    private func setupLayout() {
+        }
+    
+    
+    
+}
+
+
+
