--- SettingsScreenApplication/ViewController.swift
@@ -1,19 +1,0 @@
-//
-//  ViewController.swift
-//  SettingsScreenApplication
-//
-//  Created by Nika Semenkova on 01.02.2023.
-//
-
-import UIKit
-
-class ViewController: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        // Do any additional setup after loading the view.
-    }
-
-
-}
-
